# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMPADU Keuangan API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/keu-tahun-ajar": {
            "get": {
                "tags": [
                    "Keu Tahun Ajar"
                ],
                "summary": "Get list of academic years",
                "operationId": "164574143b5a25049fbebb0f59286165",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/KeuTahunAjar"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Keu Tahun Ajar"
                ],
                "summary": "Membuat tahun ajaran baru",
                "operationId": "2963071affe96f9542b3e0ea21bb6cfc",
                "requestBody": {
                    "description": "Data tahun ajaran",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tahun_ajaran",
                                    "semester"
                                ],
                                "properties": {
                                    "tahun_ajaran": {
                                        "type": "string",
                                        "example": "2023/2024"
                                    },
                                    "semester": {
                                        "type": "string",
                                        "example": "Ganjil"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data berhasil dibuat"
                    }
                }
            }
        },
        "/api/keu-tahun-ajar/{id}": {
            "get": {
                "tags": [
                    "Keu Tahun Ajar"
                ],
                "summary": "Mendapatkan detail tahun ajaran",
                "description": "Mengembalikan data detail tahun ajaran berdasarkan ID",
                "operationId": "getTahunAjarById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID tahun ajaran",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KeuTahunAjar"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Keu Tahun Ajar"
                ],
                "summary": "Memperbarui tahun ajaran",
                "description": "Memperbarui data tahun ajaran berdasarkan ID",
                "operationId": "updateTahunAjar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID tahun ajaran yang akan diperbarui",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data tahun ajaran yang diperbarui",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tahun_ajaran": {
                                        "type": "string",
                                        "example": "2023/2024"
                                    },
                                    "semester": {
                                        "type": "string",
                                        "example": "Ganjil"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KeuTahunAjar"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Keu Tahun Ajar"
                ],
                "summary": "Menghapus tahun ajaran",
                "description": "Menghapus data tahun ajaran berdasarkan ID",
                "operationId": "deleteTahunAjar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID tahun ajaran yang akan dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Data berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "KeuTahunAjar": {
                "required": [
                    "tahun_ajaran",
                    "semester"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tahun_ajaran": {
                        "type": "string",
                        "example": "2023/2024"
                    },
                    "semester": {
                        "type": "string",
                        "enum": [
                            "Ganjil",
                            "Genap"
                        ],
                        "example": "Ganjil"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}